requirements.txt
google-genai
python-dotenv
nltk
langgraph
pydantic[email]
scikit-learn
streamlit
langchain_groq
